Use fixed width integer cast in std::hash<std::pair>().

+++ cpplib/hash/std_pair_hash.hpp
@@ -8,7 +8,7 @@ namespace std {
 template<typename T, typename U>
 struct hash<std::pair<T, U>> {
 	std::size_t operator()(const std::pair<T, U>& pair) const {
-		return pair.first * 877117 + pair.second;
+		return static_cast<int64_t>(pair.first) * 877117 + pair.second;
 	}
 };
 #endif
